Fix `contrib`'s dependencies crashing setuptools' `extra_requires`
PyTorch is not available in PyPi requiring adjustments to requirements
file and setuptools' setup.py.
Requirement file requires to add a line of `--find-link` so pip knows
where to look for Torch (in a way that is also compatible with Travis'
`pip install -r`).
However that `--find-link` line causing setuptools' `extra_requires`
parser to crash (as it is not a valid package).

In order to avoid managing duplicated dependencies, I added a simple
functions that filters invalid lines before providing it as
`extra_require`.

This way it is possible to install requirements with both
`pip install -r contrib\requirements.txt`
and
```pip install causallib[contrib] -f
https://download.pytorch.org/whl/torch_stable.html```
(where the `-f` is needed in order to find torch, otherwise install
fails)