Fix deepcopy fail in PropensityFeatureStandardization
removed calculating of feature functions on init
this makes the object un-deepcopyable.
the reason is that the set of feature functions store a copy of self
that does not get updated on deepcopy.
since the function is only called once, and it is a light function,
i opted to remove the attribute entirely and just create the dict
when it is needed. that way self will never be misreferenced.

Add deepcopy tests