Move treatment prediction to `PropensityPredictions` in `evaluation`
Predicting the treatment label is only defined when you have a
propensity model, not a general weighting model.

Treatment label prediction is used for some of the `evaluation` metrics,
but those metrics are not well-defined for weight models.

Since the weight models can only be evaluated on their covariate balance
(currently), then it no longer need to use the compounded Result object
`WeightEvaluatorScores` and can return a covariate-balancing DataFrame
(similar to how `OutcomePredictions` only returns a DataFrame with the
scores).

Signed-off-by: Ehud-Karavani <ehud.karavani@ibm.com>