Refactor column-renaming functionality from Standardization to utils
First, this is not a Standardization-specific code, but rather
maintenance code that is used by Standardization
(which joins `a` to `X` part of its function).
Second, this might be needed elsewhere in the package.

Therefore, it is best to move this functionality to some general utils.

Signed-off-by: Ehud-Karavani <ehud.karavani@ibm.com>

Make `a_name` optional

Signed-off-by: Ehud-Karavani <ehud.karavani@ibm.com>