Fix pandas future deprecation for ignoring errors in `pd.to_numeric`
Pandas may in the near future stop supporting the capability of ignoring
columns that cannot be converted to numerics, keeping the input as is.
`FutureWarning: errors='ignore' is deprecated and will raise in a future version. Use to_numeric without passing `errors` and catch exceptions explicitly instead`

So instead, let pandas place NaNs wherever it did not succeed and then
fill those nans with the original (non-numeric) values.

Signed-off-by: Ehud-Karavani <ehud.karavani@ibm.com>