pandas version-dependent `map` or `applymap`
It seems the transition from `applymap` to `map` is <1 year old,
starting pandas version 2.1:
https://pandas.pydata.org/pandas-docs/version/2.1/reference/api/pandas.DataFrame.map.html

This is too new to fully rely on.
Especially when it breaks for python 3.7 and 3.8
that are not supported by newer pandas versions,
which I'm not yet comfortable with (although their end-of-life status,
because I'm not sure how causallib supports 3.12 at the moment).

So putting a pandas version-dependent choice makes sense at the moment
and can later be discarded.

Signed-off-by: Ehud-Karavani <ehud.karavani@ibm.com>